[skip ci] unified MESA setup - To be tested
The new inlists are obtained by systematically merging the content of
[DissipationBalancedPenetration](https://github.com/colej/DissipationBalancedPenetration/tree/main/work/inlists)

Turn on `mesa_45.net` winds, overshooting, profile output, and
evolution until C depletion. The `run_star_extras.f90` allows to
on-the-fly change the nuclear reaction network to `mesa_128.net` and
continue to the onset of core-collapse.

The `inlist_pgstar_*` have been renamed and moved to the subfolder
`pgstar` and configured to save plots in `png`.

I've modified the `rn` script to run `inlist_dbcp` if no argument is
provided, or the inlist at the path provided (like previous behavior)

If you install and activate the conda environment from the file
`environment.yaml` you will be able to check the differences between
your inlist and the ones in `src/data/MESA_input` by doing

```
compare_inlists <inlist1> <inlist2>
```
to print all the "meaningful" differences between the two
inlists (ignoring comments, order, and defaults).